New hit animation and compile correction

+++ Troubadour/core/src/com/troubadour/game/states/PlayState.java
@@ -107,7 +107,7 @@ public class PlayState extends State {
             sb.draw(player.lifeAnimation.getFrame(), cam.position.x + cam.viewportWidth - 150, cam.position.y + cam.viewportHeight - (205+20*i));
         }
         yourBitmapFontName.setColor(1.0f, 1.0f, 0f, 1.0f);
-        yourBitmapFontName.draw(sb, yourScoreName, 15, cam.position.y + cam.viewportHeight - (290));
+        yourBitmapFontName.draw(sb, yourScoreName, 15, (int) cam.position.y + cam.viewportHeight - (290));
         sb.end();
     }
